+++ src/com/deco/login/UserInfoAction.java
@@ -6,6 +6,7 @@ import javax.servlet.http.HttpSession;
 
 import com.deco.Action;
 import com.deco.ActionForward;
+import com.deco.user.userDTO;
 
 
 public class UserInfoAction implements Action{
