+++ src/com/deco/user/MailSendServlet.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import java.io.IOException;
 import java.io.PrintWriter;
