+++ src/com/deco/user/RandomCode.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import java.util.Random;
