+++ src/com/deco/user/dataController.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import java.io.IOException;
