+++ src/com/deco/user/getBody.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import java.io.BufferedReader;
 import java.io.IOException;
