+++ src/com/deco/user/sendMailAuthcode.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import javax.activation.*;
 import java.util.*;
