+++ src/com/deco/user/subjectDAO.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
