+++ src/com/deco/user/userController.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import java.io.IOException;
