+++ src/com/deco/user/userDTO.java
@@ -1,4 +1,4 @@
-package com.deco.user.join;
+package com.deco.user;
 
 import javax.servlet.http.HttpServletRequest;
